问题页面更新（增加QuizTag相关配置）

+++ Maven项目/Utunan/src/main/java/com/utunan/mapper/community/QuizMapper.java
@@ -28,5 +28,5 @@ public interface QuizMapper {
     Quiz getQuiz1();
 
     /*王碧云 通过quizId查找问题*/
-    Quiz findQuizById(@Param("quizId") int quizId);
+    Quiz findQuizById(@Param("quizId") Long quizId);
 }
